requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
RPi.GPIO==0.7.1
google-generativeai==0.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
